% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitchen-sink.R
\name{count_distinct}
\alias{count_distinct}
\title{wrapper around dplyr \code(distinct() %>% count())}
\usage{
count_distinct(x, ..., wt = NULL, sort = FALSE)
}
\arguments{
\item{x}{sthe data frame to perform over}

\item{...}{vars to perform over}

\item{wt}{wt parameter for \code(count())}

\item{sort}{sort parameter for \c0de(count())}
}
\description{
Combine \code(count()) and \code(distinct()) !
}
