% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitchen-sink.R
\name{drive_download_all}
\alias{drive_download_all}
\title{Download all files from Google Drive folder}
\usage{
drive_download_all(path, temp = TRUE, fileExt = NULL,
  saveToPath = "./", ...)
}
\arguments{
\item{path}{folder path, id, or url to pass into \code{googledrive::drive_ls}. Use \code{googledrive::as_id()} if URL or id.}

\item{temp}{boolean. Should it save to a temp location? If FALSE the files are saved in the directory specified in saveToPath.}

\item{fileExt}{list of characters, optional. only download files with specified file extensions. Do not include a ".", just the extension.}

\item{saveToPath}{character, optional. If temp = FALSE then defaults to the current directory, otherwise pass in a folder path with trailing "/"}

\item{...}{values to pass into \code{googledrive::drive_download()}}
}
\description{
Wrapper around \code{googledrive::download} to pull all files located in a Google Drive folder.
Optionally filter for files of a specific type or save to a temporary folder.
Returns a vector of file paths to the downloaded files.
}
