% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitchen-sink.R
\name{gs_large_upload}
\alias{gs_large_upload}
\title{Write a larger file to Google Sheets by writing csv then uploading that file}
\usage{
gs_large_upload(data, sheet_title, overwrite = TRUE, ...)
}
\arguments{
\item{data}{the dataframe to upload.}

\item{sheet_title}{the title of the newly uploaded sheet}

\item{overwrite}{should it overwrite a sheet of the same name? defaults to \code{TRUE}.}

\item{...}{parameters to pass into \code{googlesheets::gs_upload()}}
}
\description{
Extends \code{googlesheets::gs_upload()} by doing the write_csv and upload task for you.
If you want to modify \code{readr::wrte_csv()} do it in seperate calls.
}
