% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitchen-sink.R
\name{read_multiple_xlsx}
\alias{read_multiple_xlsx}
\title{Import a bunch of xlsx files}
\usage{
read_multiple_xlsx(search_string = ".", folderPath = "./",
  add_filename_column = FALSE, ...)
}
\arguments{
\item{search_string}{character. The string to detect using \code{stringr::str_detect} in the file name to pull in a subset of files. defaults to every file in the given folder path.}

\item{folderPath}{character. the relative path to the directory the files are in. defaults to the working directory.}

\item{add_filename_column}{boolean. Add the filename to the dataframe as a column?}

\item{...}{parameters to pass into \code{readxl::read_xlsx}}
}
\description{
This imports a set of xlsx files based on file extension and search string. It returns a list containing a dataframe for each file.
It is a simple extension of readr::read_xlsx
}
