% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitchen-sink.R
\name{tidy_acf}
\alias{tidy_acf}
\title{produces a tidy output of the \code{acf} function from \code{rstats}}
\usage{
tidy_acf(data, value, lags = 0:20)
}
\arguments{
\item{data}{dataframe or tibble}

\item{value}{the value in the datframe to produce for}

\item{lags}{the lags to use, defaults to 0:20}
}
\description{
acf function output as a tibble!
credit goes to Matt Dancho at Business Science in the time series with Keras tutorial \href{https://www.business-science.io/timeseries-analysis/2018/04/18/keras-lstm-sunspots-time-series-prediction.html}{here}
}
